Convert model to bf16 on load

+++ fork/table/transformer_rs.py
@@ -33,8 +33,7 @@ class PositionalEncoding(nn.Module):
         pe[:, 0::2] = torch.sin(position * div_term)
         pe[:, 1::2] = torch.cos(position * div_term)
         pe = pe.unsqueeze(0).transpose(0, 1)
-        # Convert to bf16 for Flash Attention compatibility
-        self.register_buffer("pe", pe.to(torch.bfloat16))
+        self.register_buffer("pe", pe)  # Keep as float32 for checkpoint compatibility
 
     def forward(self, x):
         x = x + self.pe[: x.size(0), :]
